% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.GRanges.R
\docType{methods}
\name{as.GRanges}
\alias{as.GRanges}
\alias{as.GRanges,Region-method}
\alias{as.GRanges,RegionList-method}
\title{Cast to GRranges}
\usage{
as.GRanges(x)

\S4method{as.GRanges}{Region}(x)

\S4method{as.GRanges}{RegionList}(x)
}
\arguments{
\item{x}{A \code{\link{Region}} object}
}
\value{
A \code{\link[GRanges]{GRanges}} object
}
\description{
Cast to GRranges
}
\keyword{DMRScan}
