% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRScan.R
\docType{package}
\name{DMRScan-package}
\alias{DMRScan}
\alias{DMRScan-package}
\title{DMRScan package}
\description{
To use, see example in vignette, or do example(dmrscan).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/christpa/DMRScan}
  \item Report bugs at \url{https://github.com/christpa/DMRScan/issues}
}

}
\author{
\strong{Maintainer}: CM Page \email{page.ntnu@gmail.com}

Authors:
\itemize{
  \item L Vos \email{Linda.Vos@Kreftregisteret.no}
  \item BK Andreassen \email{Bettina.Kulle.Andreassen@Kreftregisteret.no}
}

Other contributors:
\itemize{
  \item TB Rounge [contributor]
  \item HF Harbo [thesis advisor]
}

}
