\name{DMRScan.phenotypes}
\docType{data}
\alias{DMRScan.phenotypes}
\title{Dichotomous phenotype for each sample}
\description{
Dichotomous variable for each sample, giving either overweight (1) or lean (0).
}
\usage{DMRScan.phenotypes}
\format{An integer vector of same length as the number of samples (100).}
\value{An object of class integer}
\source{Trine B Rounge 2016-10-03}
\references{
    Rounge, et.al _Epigenomics_ 2016
}
\keyword{datasets}
