% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/DMRscan.R
\name{DMRscan}
\alias{DMRscan}
\title{Find differentially methylated regions using the scan statistic.}
\usage{
DMRscan(meth, mydir = NULL, xvalues = "Position", splitcentromere = FALSE,
  build = c("hg18", "hg19", "hg38"), smallestregion = 4,
  plotresult = FALSE, ...)
}
\arguments{
\item{meth}{input data, a  methylSigData object.}

\item{mydir}{a directory into which temporary files can be written.}

\item{xvalues}{defaults to "Position" to use the position values associated with the data.  Use "Index" to treat the data points as equally spaced along the chromosome.}

\item{splitcentromere}{splits data at the centromere within each chromosome.  May be necessary for large datasets.  Must also specify build.}

\item{build}{can be either hg18, hg19, or hg38.  This is necessary when splitcentromere=TRUE and/ or plotresult=TRUE.}

\item{smallestregion}{specifies the smallest region (in number of data points) that will be returned as a DMR. Default=4.}

\item{plotresult}{will output a graphical representation of each DMR to a pdf file in the specified directory.   Must specify build.}
}
\value{
a list containing matrix of DMR results, as well as other values and parameters which can be passed to plotSatScanresult.
}
\description{
Find differentially methylated regions using the scan statistic.
}
\examples{
result<-DMRscan(meth=meth, mydir=getwd(), xvalues="Index",
splitcentromere=TRUE, build="hg18", smallestregion=4, plotresult=TRUE)
}
\seealso{
\code{\link{SatScanrun}} which this function wraps, and \code{\link{plotSatScanresult}} which plots the results of this function. \code{\link{plotSatScanresult}} can be called within this function with plotresult=TRUE.
}

