\name{DMRScan.sequencingWeight}
\docType{data}
\alias{DMRScan.sequencingWeight}
\title{Sequencing weights for DMRScan.methylationData}
\description{
    Sequencing weights for each CpG in each sample, indicating sequencing depth and covarage. 
}
\usage{DMRScan.sequencingWeight}
\format{A matrix with integer-values, with individuals as columns, and genomic position as rows}
\value{An object of type matrix}
\source{Trine B Rounge 2016-10-03}
\references{
    Rounge, et.al _Epigenomics_ 2016
}
\keyword{datasets}
