% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_t_grid.R
\name{estimate_t_grid}
\alias{estimate_t_grid}
\title{Estimate window thresholds for sliding window, one unique value for each window size}
\usage{
estimate_t_grid(L, k_grid, method = "siegmund", mcmc = 1000, n_cpu = 1,
  ...)
}
\arguments{
\item{L}{The number of probes (CpGs) in the study.}

\item{k_grid}{The different window sizes to be tested. Must be either one, or an ordered sequence of integers.}

\item{method}{Gives the method by which the threshold is calculated. Can be either an analytical solution 
"siegmund", provided by Siegnumd et.al (2012), or an iterative process; either important sampling "zhang", as suggested 
by Zhang (2012) or a full MCMC model "mcmc" which can account for any dependency structure, wich is pass to arima.sim, with ...}

\item{mcmc}{The number of MCMC iterations to be used, when using either Important Sampling ("zhang") or MCMC estimation of the threshold.}

\item{n_cpu}{When calculating the thresholds on a cluster, how many CPUs should be used.}

\item{...}{Optinal parameters pased on to arima(), when simulating data using the mcmc option, see arima.sim()}
}
\value{
Returns a vector of the threshold for each window size
}
\description{
Estimate window thresholds for sliding window, one unique value for each window size
}
\examples{
t_grid <- estimate_t_grid(L = 1000, k_grid = 3:8, method = "siegmund")
}

