% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{getP}
\alias{getP}
\alias{getP,Region-method}
\alias{getP,RegionList-method}
\title{Method getP}
\usage{
getP(region)

\S4method{getP}{Region}(region)

\S4method{getP}{RegionList}(region)
}
\arguments{
\item{region}{An object of type Region or RegionList}
}
\description{
Method getP

Get p-values for a region

Get p-values for a list of regions (RegionList)
}

