% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{getT}
\alias{getT}
\alias{getT,Region-method}
\alias{getT,RegionList-method}
\title{Method getT}
\usage{
getT(region, ...)

\S4method{getT}{Region}(region, index = NULL)

\S4method{getT}{RegionList}(region, index = NULL)
}
\arguments{
\item{region}{An opbject of type Region or RegionList}

\item{...}{Index}

\item{index}{Index to extract}
}
\description{
Method getT

Get test statistic for an object of type Region

Get test statistic for all regins within a RegionList class
}

