% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{head,RegionList-method}
\alias{head,RegionList-method}
\title{Cat the head of a list of regions in a RegionList object}
\usage{
\S4method{head}{RegionList}(x, n = 10L)
}
\arguments{
\item{x}{An object to be printed of type RegionList}

\item{n}{The number of regions to be printed when the RegionList is 
longer than n}
}
\value{
The top regins in a RegionList
}
\description{
Cat the head of a list of regions in a RegionList object
}

