% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{[}
\alias{[}
\title{Get a individual region within an object of class RegionList}
\usage{
\S4method{[}{RegionList,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{An object of type RegionList}

\item{i}{Index, which region to extract}
}
\value{
A region from a RegionList of class "list"
}
\description{
Get a individual region within an object of class RegionList
}

