% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{[[}
\alias{[[}
\title{Get a region from a RegionList object}
\usage{
\S4method{[[}{RegionList}(x, i)
}
\arguments{
\item{x}{An object of type RegionList}

\item{i}{Index, which region to extract}
}
\value{
A region from a RegionList with class "Region"
}
\description{
Get a region from a RegionList object
}

