% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{initialize-Region}
\alias{initialize-Region}
\title{Constructor for class Region}
\usage{
\S4method{initialize}{Region}(.Object, tValues = "numeric",
  id = "character", position = "integer", chromosome = "character",
  pVal = "numeric")
}
\arguments{
\item{.Object}{Empty}

\item{tValues}{A vector of test statistics}

\item{id}{A vector of id's for each test statistic}

\item{position}{A vector of position for each test statistc}

\item{chromosome}{An character describing the chromosome (1-22, X,Y)}

\item{pVal}{A numeric giving the estimated p-value for each significant DMR}
}
\value{
An object of type Region
}
\description{
Constructor for class Region
}

