% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{initialize-RegionList}
\alias{initialize-RegionList}
\title{Constructor RegionList}
\usage{
\S4method{initialize}{RegionList}(.Object, nRegions = "integer",
  regions = "array")
}
\arguments{
\item{.Object}{Empty}

\item{nRegions}{The number of regions to be included in the RegionList object}

\item{regions}{An optional list of regions to be assigned to a RegionList}
}
\value{
An object of tpe RegionList
}
\description{
Constructor RegionList
}

