% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{length,Region-method}
\alias{length,Region-method}
\alias{length,RegionList-method}
\title{Calculate the length of a region in terms of CpGs}
\usage{
\S4method{length}{Region}(x)

\S4method{length}{RegionList}(x)
}
\arguments{
\item{x}{A RegionList object}
}
\value{
The number of CpGs in a Region

The number of CpGs in a RegionList
}
\description{
Calculate the length of a region in terms of CpGs

Get the number of regions in a RegionList
}
<<<<<<< HEAD
=======

>>>>>>> upstream/master
