% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.cpg.regions.R
\name{make.cpg.regions}
\alias{make.cpg.regions}
\title{Bundles CpGs into a list of clusters}
\usage{
make.cpg.regions(chr, pos, test_statistic, max.gap = 500, min.cpg = 2)
}
\arguments{
\item{chr}{(Sorted) Vector of chromosome location for each CpG (use integer annotation 1-22, X = 23, Y = 24, XY = 25, MT = 26)}

\item{pos}{(Sorted) Vector giving base pair position for each CpG 
If unsorted, use order(chr,pos) to sort the genomic positions within each chromosome.}

\item{test_statistic}{Vector of corresponding observed T-value for each CpG, must be ordered in the same way as chr and pos}

\item{max.gap}{Maximum allowed base pair gap within a cluster. Default is set to 500.}

\item{min.cpg}{Minimum number of CpGs allowed in each region to be considered. Default is set to at least 2 CpGs within each region.}
}
\value{
The suplied test_statistics ordered into into a list, with one entry for each CpG region.
}
\description{
This function creates a list of regions/clusters of CpGs in close proximity, with no gap larger than max.gap, and 
all where all regions are longer than min.cpg.
}
\examples{
data(DMRScan.methylationData) ## Load methylation data from chromosome 22
data(DMRScan.phenotypes) ## Load phenotype (end-point for methylation data)

## Test for an association between phenotype and Methylation
test.statistics <- apply(DMRScan.methylationData,1,function(x,y)summary(glm(y ~ x, family = binomial(link = "logit")))$coefficients[2,3], y = DMRScan.phenotypes)
pos<- data.frame(matrix(as.integer(unlist(strsplit(names(test.statistics),split="chr|[.]"))), ncol = 3, byrow = TRUE))[,-1] ## Set chromosomal position to each test-statistic

## Set clustering features 
min.cpg <- 3  ## Minimum number of CpGs in a tested cluster
max.gap <- 750  ## Maxium distance (in base-pairs) within a cluster before it is broken up into two seperate cluster 
regions <- make.cpg.regions(test_statistic = test.statistics, chr = pos[,1], pos = pos[,2], max.gap = max.gap, min.cpg = min.cpg)
}
\keyword{CpG}
\keyword{Regions}

