% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{names,Region-method}
\alias{names,Region-method}
\alias{names,RegionList-method}
\title{Get the names of all probes within a region}
\usage{
\S4method{names}{Region}(x)

\S4method{names}{RegionList}(x)
}
\arguments{
\item{x}{An object of type Region}
}
\value{
The names of individual CpGs in a Region

A character vector of all CpG ids in a RegionList
}
\description{
Get the names of all probes within a region

Get the names of all probes in a study
}

