% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{nCpG}
\alias{nCpG}
\alias{nCpG,Region-method}
\alias{nCpG,RegionList-method}
\title{Method nCpG}
\usage{
nCpG(x)

\S4method{nCpG}{Region}(x)

\S4method{nCpG}{RegionList}(x)
}
\arguments{
\item{x}{An opbject of type Region or RegionList}
}
\value{
The number of CpGs in an object
}
\description{
Method nCpG

Get the number of CpGs i a region

Get the number of CpGs in a RegionList
}
\examples{
someEmptyRegions <- RegionList(3L)
# The number of CpGs in this regions is 0
nCpG(someEmptyRegions)

}

