% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegion.R
\name{plot.Region}
\alias{plot}
\alias{plot.Region}
\title{Plot DMRs of type Region}
\usage{
\method{plot}{Region}(x, ...)
}
\arguments{
\item{x}{A Region object to be ploted. Can be subsetted from RegionList}

\item{...}{Inherited from plot()}
}
\description{
Plot DMRs of type Region
}

