% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{pos}
\alias{pos}
\alias{pos,Region-method}
\alias{pos,RegionList-method}
\title{Method pos}
\usage{
pos(region)

\S4method{pos}{Region}(region)

\S4method{pos}{RegionList}(region)
}
\arguments{
\item{region}{An opbject of type Region or RegionList}
}
\value{
An integer vector of positions for each probe site
}
\description{
Method pos

Get the chromosomal coordinates for a Region

Get the chromosomal coordinates for a 
list of regions in a RegionList object
}
\examples{
#Number of probes is n = 10
nCpG <- 10
region <- Region(tValues    = rnorm(nCpG),
                 position   = 1:nCpG,
                 chromosome = "3")
## Genomic coordinates for Region
pos(region)
}

