% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{pVal}
\alias{pVal}
\alias{pVal,Region-method}
\alias{pVal,RegionList-method}
\title{Method get pvalue}
\usage{
pVal(region, n = 12)

\S4method{pVal}{Region}(region, n = 12)

\S4method{pVal}{RegionList}(region, n = 12)
}
\arguments{
\item{region}{An object of type Region or RegionList}

\item{n}{The number of digits to be presented. Default is 10}
}
\value{
The reported p-value for a region
}
\description{
Method get pvalue

Get p-values for a region

Get p-values for a list of regions (RegionList)
}
\examples{
#Number of probes is n = 10
nCpG <- 10
region <- Region(tValues    = rnorm(nCpG),
                 position   = 1:nCpG,
                 chromosome = "3",
                 pVal       = runif(1))
## Pvalues for Region
pVal(region)
}
