% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{range,Region-method}
\alias{range,Region-method}
\title{Get the genomic position of a Region}
\usage{
\S4method{range}{Region}(x)
}
\arguments{
\item{x}{An object of type Region}
}
\value{
A character giving the genomic position
}
\description{
Get the genomic position of a Region
}
