% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\name{Region-init}
\alias{Region}
\alias{Region-init}
\title{Shorthand for initializing region}
\usage{
Region(...)
}
\arguments{
\item{...}{Parameters to pased to new("Regions",...)}
}
\value{
An object of type Region
}
\description{
Shorthand for initializing region
}
\examples{
#Number of probes is n = 10
nCpG <- 10
region <- Region(tValues    = rnorm(nCpG),
                 position   = 1:nCpG,
                 chromosome = "3")
}

