% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/SatScanrun.R
\name{SatScanrun}
\alias{SatScanrun}
\title{Find differntially methylated regions using the scan statistic.}
\usage{
SatScanrun(meth, xvalues = c("Index", "Position"), smallestregion,
  dir = tmpdir)
}
\arguments{
\item{meth}{input data, a  methylSigData object.}

\item{xvalues}{either "Index" to use the position values associated with the data, or "Index" to treat the data points as equally spaced along the chromosome.}

\item{smallestregion}{specifies the smallest region (in number of data points) that will be returned as a DMR. Default=4.}

\item{dir}{is the temporary directory where SatScan files will be written.}

\item{plotresult}{will output a graphical representation of each DMR to a pdf file in the specified directory.   Must specify build.}
}
\value{
a list containing matrix of DMR results, as well as the likelihood values needed for plotting.
}
\description{
Find differntially methylated regions using the scan statistic.
}
\examples{
SatScanrun()
}
\seealso{
\code{\link{methdiffSatScan}} wraps this function, and allows its use on data with multiple chromosomes.
}

