% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{setRegion}
\alias{setRegion}
\alias{setRegion,RegionList-method}
\title{Method setRegion}
\usage{
setRegion(x, i, ...)

\S4method{setRegion}{RegionList}(x, i, region)
}
\arguments{
\item{x}{A region}

\item{i}{an index}

\item{...}{To be pased to Region()}

\item{region}{An object of type Region to be inseted in RegionList}
}
\value{
An updated version of RegionList x, with a new Region at index i
}
\description{
Method setRegion

Update a RegionList object
}
\examples{
## A region list with 3 regions
regList <- RegionList(3L)
#Number of probes in first is n = 10
nCpG <- 10
region <- Region(tValues    = rnorm(nCpG),
                 position   = 1:nCpG,
                 chromosome = "3")
## Set first region in regList to region
regList <- setRegion(regList,i = 1, region)

}
<<<<<<< HEAD
=======

>>>>>>> upstream/master
