% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{show,Region-method}
\alias{show,Region-method}
\title{Show a region}
\usage{
\S4method{show}{Region}(object)
}
\arguments{
\item{object}{The region to be desplied, of type Region}
}
\description{
Show a region
}

