% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{sort,RegionList-method}
\alias{sort,RegionList-method}
\title{Sort a set of regions on p-value in a RegionList object}
\usage{
\S4method{sort}{RegionList}(x, decreasing = FALSE)
}
\arguments{
\item{x}{An object of type RegionList}

\item{decreasing}{Inherited from base}
}
\value{
An updated RegionList, sorted on empirical p-values
}
\description{
Sort a set of regions on p-value in a RegionList object
}

