% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{tVal}
\alias{tVal}
\alias{tVal,Region-method}
\alias{tVal,RegionList-method}
\title{Method get T statistic for a region}
\usage{
tVal(region, ...)

\S4method{tVal}{Region}(region, index = NULL)

\S4method{tVal}{RegionList}(region, index = NULL)
}
\arguments{
\item{region}{An opbject of type Region or RegionList}

\item{...}{Index}

\item{index}{Index to extract}
}
\value{
A numeric vector of t-values for a Region or RegionList
}
\description{
Method get T statistic for a region

Get test statistic for an object of type Region

Get test statistic for all regins within a RegionList class
}
\examples{
#Number of probes is n = 10
nCpG <- 10
region <- Region(tValues    = rnorm(nCpG),
                 position   = 1:nCpG,
                 chromosome = "3")
## T values for Region
tVal(region)

}
