\name{DMRScan.methylationData}
\docType{data}
\alias{DMRScan.methylationData}
\title{RRBS data on Chromosome 22 on 100 Finish teens.}
\description{
    RRBS data from chromosome 22 in 100 extreme BMI (50 high, 50 low), 
    with row names indicating the genomic coordinate (GRCh37).
}
\usage{DMRScan.methylationData}
\format{A matrix with methylation _beta_-values, with individuals as columns, and genomic position as rows}
\value{An object of type matrix}
\source{Trine B Rounge 2016-10-03}
\references{
    Rounge, et.al _Epigenomics_ 2016
}
\keyword{datasets}
