% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRScan.methylationData.R
\docType{data}
\name{DMRScan.methylationData}
\alias{DMRScan.methylationData}
\title{DMRScan}
\description{
Bi-sulfite sequencing data from all known CpG islands  at chromosome 22 
from 100 the Finish teens study, sampled from extreme BMI quantiles. The data set is 
reduced to 25139 sites on chromosome 22. 
See "Genome-wide DNA methylation in saliva and body size of adolescent girls", TB Rounge, CM Page, 
M Lepisto, E Pekka, and BK Andreassen and E Weiderpass, _Epigenomics_ 8.11 (2016): 1495-1505 for a 
full overview of the data set.
}
\examples{
data(DMRScan.methylationData)
head(DMRScan.methylationData)
}
\keyword{datasets}
