% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRScan.phenotypes.R
\docType{data}
\name{DMRScan.phenotypes}
\alias{DMRScan.phenotypes}
\title{DMRScan}
\description{
Phenotypes for methylation data, indicating case control status. 
See "Genome-wide DNA methylation in saliva and body size of adolescent girls", 
TB Rounge, CM Page, M Lepisto, E Pekka, and BK Andreassen and E Weiderpass, 
_Epigenomics_ 8.11 (2016): 1495-1505 for a full description of the phenotypes.
}
\examples{
data(DMRScan.phenotypes)
table(DMRScan.phenotypes)
}
\keyword{dataset}
