% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pakage-name.R
\docType{package}
\name{DMRScan}
\alias{DMRScan}
\alias{DMRScan-package}
\title{DMRScan: An R-package for identification of Differentially Metylated Regions}
\description{
DMRScan: An R-package for identification of Differentially Metylated Regions
}
\author{
Christian Page, \email{page.ntnu@gmail.com}
}
\references{
\url{http://Some_link_to_BMC-bioInfomatics.com}
}
\keyword{DMR}

