% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRScan.R
\name{dmrscan}
\alias{dmrscan}
\title{DMR Scan function}
\usage{
dmrscan(obs, k_grid, t_grid = NULL, ...)
}
\arguments{
\item{obs}{[numeric] Sequence of test statistics for each CpG}

\item{k_grid}{[integer] Sequence of window sizes for the sliding window}

\item{t_grid}{[numeric] Optional argument with corresponding cut-off for each window. Will be estimated if not supplied}
}
\description{
This function search for DMRs given a orded sequence of test statistics
}
\examples{

L               <- 20000 # Number of probes
pos             <- cumsum(rpois(L,30))
chr             <- rep(1,L) ## use plink annotation 1-22, X = 23, Y = 24, XY = 25, MT = 26
spike           <- numeric(L); spike[(L/2 - 10):(L/2 + 10)] <- 2 #Add i DMRs around L/2
z_val           <- matrix(rnorm(L, mean = spike),dimnames=list(paste("chr1",pos, sep = ".")))

min.cpg <- 3  # minimum number of cps for a region to ba chnsidered
max.gap <- 30 # Maximum allowd gap between two observations within one region
regions <- make.cpg.regions(test_statistic=z_val,chr=chr,pos=pos,max.gap=max.gap,min.cpg=min.cpg)

nobs    <- sum(sapply(regions,nrow)) ## observations remaining

k_grid  <- c(2,3,4) ## window sizes 
mcmc    <- 1000 ## number of simulation for window sizes to be deterimed

t_grid  <- estimate_t_grid(k_grid=k_grid,L=L,method = "zhang",mcmc=mcmc)
res     <- dmrscan(obs=regions,k_grid=k_grid,t_grid=t_grid) ## If no regions are found, the function returns NA

}
\keyword{DMRScan}

