% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_t_grid.R
\name{estimate_t_grid}
\alias{estimate_t_grid}
\title{Estimate window thresholds for sliding window, one unique value for each window size}
\usage{
estimate_t_grid(L, k_grid, method = "siegmund", ...)
}
\arguments{
\item{L}{[integer] The number of probes in the study.}

\item{k_grid}{[integer] The different window sizes to be tested. Must be an ordered sequence of integers}

\item{method}{[string] Gives the method by which the threshold is calculated. Can be either an analytical solution "siegmund", provided by Siegnumd et.al (2012), or an iterative process; either important sampling "zhang", as suggested by Zhang (2012) or a full MCMC model "mcmc" which can account for any dependency structure, passed to the argument "foo".
Optional Parameters}

\item{...}{Optinal parameters pased on to arima(), when simulating data.}

\item{mcmc}{[integer] The number of MCMC iterations to be used, when using either Important Sampling ("zhang") or MCMC estimation of the threshold.}

\item{n_cpu}{[integer] When calculating the thresholds on a cluster, how many CPUs should be used.}

\item{foo}{[string] Function call for the generation of data in the MCMC simulation, can be either "ar", "ma", or "arima".}
}
\value{
Returns a vector of the threshold for each window size
}
\description{
Estimate window thresholds for sliding window, one unique value for each window size
}
\examples{

}

