% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{getP}
\alias{getP}
\alias{getP,Region-method}
\alias{getP,RegionList-method}
\title{Method getP}
\usage{
getP(region, n = 10)

\S4method{getP}{Region}(region, n = 10)

\S4method{getP}{RegionList}(region, n = 10)
}
\arguments{
\item{region}{An object of type Region or RegionList}

\item{n}{The number of digits to be presented. Default is 10}
}
\value{
A numeric vector of p-values
}
\description{
Method getP

Get p-values for a region

Get p-values for a list of regions (RegionList)
}
\examples{
#Number of probes is n = 10
nCpG <- 10
region <- Region(tValues    = rnorm(nCpG),
                 position   = 1:nCpG,
                 chromosome = "3",
                 pVal       = runif(1))
## Pvalues for Region
getP(region)
}

