% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{getPos}
\alias{getPos}
\alias{getPos,Region-method}
\alias{getPos,RegionList-method}
\title{Method getPos}
\usage{
getPos(region)

\S4method{getPos}{Region}(region)

\S4method{getPos}{RegionList}(region)
}
\arguments{
\item{region}{An opbject of type Region or RegionList}
}
\value{
An integer vector of positions for each probe site
}
\description{
Method getPos

Get the chromosomal coordinates for a Region

Get the chromosomal coordinates for a 
list of regions in a RegionList object
}
\examples{
#Number of probes is n = 10
nCpG <- 10
region <- RegionInit(tValues    = rnorm(nCpG),
                 position   = 1:nCpG,
                 chromosome = "3")
## Genomic coordinates for Region
getPos(region)
}

