% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{getRegions}
\alias{getRegions}
\alias{getRegions}
\alias{getRegions,RegionList-method}
\title{Method getRegions}
\usage{
getRegions(x)
}
\arguments{
\item{x}{An object of type RegionList}
}
\value{
An object of type Region

A region from a RegionList
}
\description{
Method getRegions

getRegions for Region List
}
\examples{
someEmptyRegions <- RegionList(3L) 
# To get back three empty regions
getRegions(someEmptyRegions)
}
