% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{getRegions}
\alias{getRegions}
\alias{getRegions,RegionList-method}
\title{Method getRegions}
\usage{
getRegions(x)

\S4method{getRegions}{RegionList}(x)
}
\arguments{
\item{x}{A RegionList object}
}
\value{
An object of type Region
}
\description{
Method getRegions

Get the regions within an RegionList object
}
\examples{
someEmptyRegions <- RegionListInit(3L) 
# To get back three empty regions
getRegions(someEmptyRegions)

}

