% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{[<-}
\alias{[}
\alias{[<-}
\title{Get a individual region within an object of class RegionList}
\usage{
\S4method{[}{RegionList,ANY}(x, i)
}
\arguments{
\item{x}{An object of type RegionList}

\item{i}{Index, which region to extract}

\item{j}{Not used}
}
\description{
Get a individual region within an object of class RegionList
}

