% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,RegionList-method}
\title{Get Object Region}
\arguments{
\item{x}{An object of type RegionList}

\item{i}{Index, which region to extract}

\item{j}{(Not used)}

\item{...}{(not used)}

\item{drop}{If drop is used}
}
\value{
A region from a RegionList of class "list"
}
\description{
Get Object Region
}

