% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{[[}
\alias{RegionList-method}
\alias{[[}
\alias{[[,}
\title{Get a region from a RegionList object}
\usage{
\S4method{[[}{RegionList}(x, i, j)
}
\arguments{
\item{x}{An object of type RegionList}

\item{i}{Index, which region to extract}

\item{j}{Not Used}
}
\description{
Get a region from a RegionList object
}

