% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.cpg.regions.R
\name{make.cpg.regions}
\alias{make.cpg.regions}
\title{Bundles CpGs into a list of clusters}
\usage{
make.cpg.regions(chr, pos, test_statistic, max.gap = 500, min.cpg = 2)
}
\arguments{
\item{chr}{(Sorted) Vector of chromosome location for each CpG (use integer annotation 1-22, X = 23, Y = 24, XY = 25, MT = 26)}

\item{pos}{(Sorted) Vector giving base pair position for each CpG 
If unsorted, use order(chr,pos) to sort the genomic positions within each chromosome.}

\item{test_statistic}{Vector of corresponding observed T-value for each CpG, must be ordered in the same way as chr and pos}

\item{max.gap}{Maximum allowed base pair gap within a cluster. Default is set to 500.}

\item{min.cpg}{Minimum number of CpGs allowed in each region to be considered. Default is set to at least 2 CpGs within each region}
}
\description{
This function creates a list of regions/clusters of CpGs in close proximity, with no gap larger than max.gap, and 
all where all regions are longer than min.cpg.
}
\examples{
pos             <- cumsum(rpois(100,30))
chr             <- sort(sample(1:3,100,TRUE)) 
z_val           <- matrix(rnorm(100),dimnames=list(paste(chr,pos, sep = ".")))

min.cpg <- 3  # minimum number of cps for a region to be chnsidered
max.gap <- 30 # Maximum allowd gap between two observations within one region
regions <- make.cpg.regions(test_statistic=z_val,chr=chr,pos=pos,max.gap=max.gap,min.cpg=min.cpg)
## Remaining observations 
sum(sapply(regions,nrow))
}
\keyword{CpG}
\keyword{Regions}

