% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{print,Region-method}
\alias{print,Region-method}
\alias{print,RegionList-method}
\title{Print a region}
\usage{
\S4method{print}{Region}(x, ...)

\S4method{print}{RegionList}(x)
}
\arguments{
\item{x}{Object of type Region}

\item{...}{Has no function}
}
\value{
An print object of a Region class

A printed object of all regions in a RegionList
}
\description{
Print a region

Print a number of regions in a RegionList
}
