% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{class}
\name{Region-class}
\alias{Region-class}
\title{Object of type Region}
\description{
Class \code{Region} is a collection of test statistics for a set of CpGs 
within a short genomic range
}
<<<<<<< HEAD
=======

>>>>>>> upstream/master
