% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\name{Region}
\alias{Region}
\title{Shorthand for initializing region}
\usage{
Region(tValues, position, chromosome, pVal, id)
}
\arguments{
\item{tValues}{A vector of test statistics}

\item{position}{A vector of position for each test statistc}

\item{chromosome}{An character describing the chromosome (1-22, X,Y)}

\item{pVal}{The P value of a region, set to numeric() if not given.}

\item{id}{The names of each probe in the region}
}
\value{
An object of type Region

An object of type Region
}
\description{
Shorthand for initializing region
}
\examples{
#Number of probes is n = 10
nCpG <- 10
region <- Region(tValues    = rnorm(nCpG),
                 position   = 1:nCpG,
                 chromosome = "3",
                 id         = paste("CpG",1:nCpG,sep="_"),
                 pVal       = runif(1)) 
}

