% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{class}
\name{RegionList-class}
\alias{RegionList-class}
\title{Class RegionList                                                        
                                                                    
Class \code{RegionList} is a collection of Regions}
\description{
Class RegionList                                                        
                                                                    
Class \code{RegionList} is a collection of Regions
}

