% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\name{RegionList}
\alias{RegionList}
\title{Shorthand for initializing RegionList}
\usage{
RegionList(nRegions, regions)
}
\arguments{
\item{nRegions}{The number of regions to be placed}

\item{regions}{The regions to be included}
}
\value{
An object of type RegionList
}
\description{
Shorthand for initializing RegionList
}
\examples{
# An empty list of 3 regions
RegionList(3L)

}
