% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R, R/test.R
\docType{methods}
\name{oneWindowSizeScanner}
\alias{Rt}
\alias{Rt,Region-method}
\alias{Rt,RegionList-method}
\alias{oneWindowSizeScanner}
\alias{oneWindowSizeScanner,Region-method}
\alias{oneWindowSizeScanner,RegionList-method}
\title{Method Fixed window size scan for one window size}
\usage{
oneWindowSizeScanner(region, windowThreshold, windowSize)

\S4method{oneWindowSizeScanner}{RegionList}(region, windowThreshold, windowSize)

\S4method{oneWindowSizeScanner}{Region}(region, windowThreshold, windowSize)
}
\arguments{
\item{region}{Object of type Region or RegionList}

\item{windowThreshold}{Vector of window thresholds}

\item{windowSize}{Vector of window sizes to be tested on regions}
}
\value{
A list of which windows that are significant
}
\description{
Method Fixed window size scan for one window size
}
\examples{
## Not run

}

