% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R, R/test.R
\docType{methods}
\name{Rt}
\alias{Rt}
\alias{Rt,Region-method}
\alias{Rt,RegionList-method}
\title{Method Rt}
\usage{
Rt(region, windowThreshold, windowSize)

\S4method{Rt}{RegionList}(region, windowThreshold, windowSize)

\S4method{Rt}{Region}(region, windowThreshold, windowSize)
}
\arguments{
\item{region}{Object of type Region or RegionList}

\item{windowThreshold}{Vector of window thresholds}

\item{windowSize}{Vector of window sizes to be tested on regions}
}
\description{
Method Rt
}

