% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R, R/test.R
\docType{methods}
\name{oneWindowSizeScanner}
\alias{oneWindowSizeScanner}
\alias{Rt}
\alias{oneWindowSizeScanner,GRangesList-method}
\alias{Rt,GRangesList-method}
\alias{oneWindowSizeScanner,GRanges-method}
\alias{Rt,GRanges-method}
\title{Method Fixed window size scan for one window size}
\usage{
oneWindowSizeScanner(region, windowThreshold, windowSize)

\S4method{oneWindowSizeScanner}{GRangesList}(region, windowThreshold,
  windowSize)

\S4method{oneWindowSizeScanner}{GRanges}(region, windowThreshold, windowSize)
}
\arguments{
\item{region}{Object of type GRanges}

\item{windowThreshold}{Vector of window thresholds}

\item{windowSize}{Vector of window sizes to be tested on regions}
}
\value{
A list of the windows that are significant
}
\description{
Method Fixed window size scan for one window size
}
\examples{
## Not run

}
