% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{setRegion}
\alias{setRegion}
\alias{setRegion,RegionList-method}
\alias{setRegion,missing-method}
\title{Method setRegion}
\usage{
setRegion(x, i, ...)

\S4method{setRegion}{RegionList}(x, i, region)

\S4method{setRegion}{missing}(x, i, ...)
}
\arguments{
\item{x}{A region}

\item{i}{an index}

\item{...}{Arguments to be pased to Region(...)}

\item{region}{An object of type Region to be inseted in RegionList}
}
\description{
Method setRegion

Update a RegionList object

Set a Region of type missing
}

