% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R, R/test.R
\docType{methods}
\name{manyWindowSizeScanner}
<<<<<<< HEAD
\alias{manyWindowSizeScanner}
\alias{St}
\alias{manyWindowSizeScanner,RegionList-method}
\alias{manyWindowSizeScanner,Region-method}
=======
\alias{St}
\alias{manyWindowSizeScanner}
\alias{manyWindowSizeScanner,Region-method}
\alias{manyWindowSizeScanner,RegionList-method}
>>>>>>> upstream/master
\title{Method Fixed window size scan for a sequence of window sizes}
\usage{
manyWindowSizeScanner(region, windowThreshold, windowSize)

\S4method{manyWindowSizeScanner}{RegionList}(region, windowThreshold,
  windowSize)

\S4method{manyWindowSizeScanner}{Region}(region, windowThreshold, windowSize)
}
\arguments{
\item{region}{Object of type Region or RegionList}

\item{windowThreshold}{Vector of window thresholds}

\item{windowSize}{Vector of window sizes to be tested on regions}
}
\value{
<<<<<<< HEAD
A list of the windows that are significant
=======
A list of which windows that are significant
>>>>>>> upstream/master
}
\description{
Method Fixed window size scan for a sequence of window sizes
}
\examples{
## Not run

}
<<<<<<< HEAD
=======

>>>>>>> upstream/master
