% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R, R/test.R
\docType{methods}
\name{manyWindowSizeScanner}
\alias{manyWindowSizeScanner}
\alias{St}
\alias{manyWindowSizeScanner,GRangesList-method}
\alias{manyWindowSizeScanner,GRanges-method}
\title{Method Fixed window size scan for a sequence of window sizes}
\usage{
manyWindowSizeScanner(region, windowThreshold, windowSize)

\S4method{manyWindowSizeScanner}{GRangesList}(region, windowThreshold,
  windowSize)

\S4method{manyWindowSizeScanner}{GRanges}(region, windowThreshold, windowSize)
}
\arguments{
\item{region}{Object of type GRanges}

\item{windowThreshold}{Vector of window thresholds}

\item{windowSize}{Vector of window sizes to be tested on regions}
}
\value{
A list of the windows that are significant
}
\description{
Method Fixed window size scan for a sequence of window sizes
}
\examples{
## Not run

}
