% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R, R/test.R
\docType{methods}
\name{St}
\alias{St}
\alias{St,Region-method}
\alias{St,RegionList-method}
\title{Method St}
\usage{
St(region, windowThreshold, windowSize)

\S4method{St}{RegionList}(region, windowThreshold, windowSize)

\S4method{St}{Region}(region, windowThreshold, windowSize)
}
\arguments{
\item{region}{Object of type Region or RegionList}

\item{windowThreshold}{Vector of window thresholds}

\item{windowSize}{Vector of window sizes to be tested on regions}
}
\value{
A list of which windows that are significant
}
\description{
Method St
}
\examples{
## Not run

}

